+++ test/vmtests/vmtests/imagecustomizer/conftest.py
@@ -1,13 +1,10 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 
-import shutil
 from pathlib import Path
 from typing import Generator
 
 import pytest
-
-from ..conftest import create_temp_folder
 
 
 def pytest_addoption(parser: pytest.Parser) -> None:
@@ -15,15 +12,6 @@ def pytest_addoption(parser: pytest.Parser) -> None:
     parser.addoption("--core-efi-azl3", action="store", help="Path to input image")
     parser.addoption("--core-legacy-azl2", action="store", help="Path to input image")
     parser.addoption("--core-legacy-azl3", action="store", help="Path to input image")
-
-
-@pytest.fixture(scope="session")
-def session_temp_dir(request: pytest.FixtureRequest, keep_environment: bool) -> Generator[Path, None, None]:
-    temp_path = create_temp_folder("vmtests-")
-    yield Path(temp_path)
-
-    if not keep_environment:
-        shutil.rmtree(temp_path)
 
 
 @pytest.fixture(scope="session")
