+++ test/vmtests/vmtests/osmodifier/conftest.py
@@ -1,27 +1,15 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 
-import shutil
 from pathlib import Path
 from typing import Generator
 
 import pytest
 
-from ..conftest import create_temp_folder
-
 
 def pytest_addoption(parser: pytest.Parser) -> None:
     parser.addoption("--input-image", action="store", help="Path to input image")
     parser.addoption("--osmodifier-binary", action="store", help="Path to osmodifier binary")
-
-
-@pytest.fixture(scope="session")
-def session_temp_dir(request: pytest.FixtureRequest, keep_environment: bool) -> Generator[Path, None, None]:
-    temp_path = create_temp_folder("osmodifiertest-")
-    yield Path(temp_path)
-
-    if not keep_environment:
-        shutil.rmtree(temp_path)
 
 
 @pytest.fixture(scope="session")
